Realocated and replaced components
Components where replaced because of the chip shortage, such as the 3.3 V Voltage regulator, the USB ESD protector and the SD Card holder.

The ESP32 has been rotated and moved closer to the center of the board, now having it's antenna positioned outside of the original board outline, providing better range for the Bluetooth and Wi-Fi.