Reduces board length to 101,5 mm
JLC PCB allows a 4 layer PCB to be manufactured for just $8,00 if it is within 102 x 102 mm. The previous version was 105,47 mm, making it cost $37,00.

To reduce the manufacturing cost the board had its size cut down to 101,5 mm. This was possible by moving the buzzer from the top left corner of the board to the free space between the two SMA connectors.